+++ analysis/Cargo.toml
@@ -13,6 +13,7 @@ syn = { version = "1.0", features = [ "full", "parsing" ] }
 derive_more = "0.99"
 prusti-rustc-interface = { path = "../prusti-rustc-interface" }
 tracing = { path = "../tracing" }
+prusti-utils = { path = "../prusti-utils" }
 
 [dev-dependencies]
 compiletest_rs = "0.10"
