+++ analysis/tests/test_analysis.rs
@@ -7,6 +7,7 @@ mod utils;
 use compiletest_rs as compiletest;
 use std::{env, path::PathBuf};
 use utils::*;
+use prusti_utils::utils::find_compiled_executable;
 
 fn run_tests(mode: &str, path: &str, custom_args: Vec<String>) {
     let mut config = compiletest::Config::default();
